Grpc/grpc-tutorial-python/requirements.txt
grpcio==1.64.1
grpcio-reflection==1.64.1
protobuf==5.27.1
python-dotenv==1.0.1
